Rename medication to take file
